Split more code from the TradeClient.
Fixed the map initialization code.
