+++ src/Sidekick.Business/Maps/IMapService.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Sidekick.Business.Maps
+{
+    public interface IMapService
+    {
+        HashSet<string> MapNames { get; }
+    }
+}
