+++ src/Sidekick.Business/StartupExtensions.cs
@@ -4,6 +4,7 @@ using Sidekick.Business.Categories;
 using Sidekick.Business.Http;
 using Sidekick.Business.Languages;
 using Sidekick.Business.Leagues;
+using Sidekick.Business.Maps;
 using Sidekick.Business.Parsers;
 using Sidekick.Business.Tokenizers;
 using Sidekick.Business.Tokenizers.ItemName;
@@ -21,6 +22,7 @@ namespace Sidekick.Business
             services.AddSingleton<IItemParser, ItemParser>();
             services.AddSingleton<ILanguageProvider, LanguageProvider>();
             services.AddSingleton<ILeagueService, LeagueService>();
+            services.AddSingleton<IMapService, MapService>();
             services.AddSingleton<IPoeApiService, PoeApiService>();
             services.AddSingleton<IStaticItemCategoryService, StaticItemCategoryService>();
             services.AddSingleton<ITokenizer, ItemNameTokenizer>();
